move non-scenario tests away from e2e module
