+++ e2e/components/components_test.go
@@ -1,4 +1,4 @@
-package e2e
+package components
 
 import (
 	"fmt"
@@ -15,7 +15,7 @@ func TestVersion1_32IsCached(t *testing.T) {
 }
 
 func TestImagesAreFullySpecified(t *testing.T) {
-	images := getWindowsContainerImages("mcr.microsoft.com/windows/servercore:*", "2025-gen2")
+	images := GetWindowsContainerImages("mcr.microsoft.com/windows/servercore:*", "2025-gen2")
 	tags := getWindowsContainerImageTags("mcr.microsoft.com/windows/servercore:*", "2025-gen2")
 	image := images[0]
 	tag := tags[0]
@@ -29,79 +29,79 @@ func TestWs2025Gen2ServerCore(t *testing.T) {
 }
 
 func TestWs2025Gen2Nanoserver(t *testing.T) {
-	serverCoreVersions := getWindowsContainerImages("mcr.microsoft.com/windows/nanoserver:*", "2025-gen2")
+	serverCoreVersions := GetWindowsContainerImages("mcr.microsoft.com/windows/nanoserver:*", "2025-gen2")
 	t.Logf("found servercore version %v", serverCoreVersions)
 	require.Len(t, serverCoreVersions, 2)
 }
 
 func TestWs2025ServerCore(t *testing.T) {
-	serverCoreVersions := getWindowsContainerImages("mcr.microsoft.com/windows/servercore:*", "2025")
+	serverCoreVersions := GetWindowsContainerImages("mcr.microsoft.com/windows/servercore:*", "2025")
 	t.Logf("found servercore version %v", serverCoreVersions)
 	require.Len(t, serverCoreVersions, 4)
 }
 
 func TestWs2025Nanoserver(t *testing.T) {
-	serverCoreVersions := getWindowsContainerImages("mcr.microsoft.com/windows/nanoserver:*", "2025")
+	serverCoreVersions := GetWindowsContainerImages("mcr.microsoft.com/windows/nanoserver:*", "2025")
 	t.Logf("found servercore version %v", serverCoreVersions)
 	require.Len(t, serverCoreVersions, 2)
 }
 
 func TestWs23H2ServerCore(t *testing.T) {
-	serverCoreVersions := getWindowsContainerImages("mcr.microsoft.com/windows/servercore:*", "23H2")
+	serverCoreVersions := GetWindowsContainerImages("mcr.microsoft.com/windows/servercore:*", "23H2")
 	t.Logf("found servercore version %v", serverCoreVersions)
 	require.Len(t, serverCoreVersions, 2)
 }
 
 func TestWs23H2Nanoserver(t *testing.T) {
-	serverCoreVersions := getWindowsContainerImages("mcr.microsoft.com/windows/nanoserver:*", "23H2")
+	serverCoreVersions := GetWindowsContainerImages("mcr.microsoft.com/windows/nanoserver:*", "23H2")
 	t.Logf("found servercore version %v", serverCoreVersions)
 	require.Len(t, serverCoreVersions, 1)
 }
 
 func TestWs23H2ServerCoreGen2(t *testing.T) {
-	serverCoreVersions := getWindowsContainerImages("mcr.microsoft.com/windows/servercore:*", "23H2-gen2")
+	serverCoreVersions := GetWindowsContainerImages("mcr.microsoft.com/windows/servercore:*", "23H2-gen2")
 	t.Logf("found servercore version %v", serverCoreVersions)
 	require.Len(t, serverCoreVersions, 2)
 }
 
 func TestWs23H2NanoserverGen2(t *testing.T) {
-	serverCoreVersions := getWindowsContainerImages("mcr.microsoft.com/windows/nanoserver:*", "23H2-gen2")
+	serverCoreVersions := GetWindowsContainerImages("mcr.microsoft.com/windows/nanoserver:*", "23H2-gen2")
 	t.Logf("found servercore version %v", serverCoreVersions)
 	require.Len(t, serverCoreVersions, 1)
 }
 
 func TestWs2022ServerCore(t *testing.T) {
-	serverCoreVersions := getWindowsContainerImages("mcr.microsoft.com/windows/servercore:*", "2022-containerd")
+	serverCoreVersions := GetWindowsContainerImages("mcr.microsoft.com/windows/servercore:*", "2022-containerd")
 	t.Logf("found servercore version %v", serverCoreVersions)
 	require.Len(t, serverCoreVersions, 2)
 }
 
 func TestWs2022Nanoserver(t *testing.T) {
-	serverCoreVersions := getWindowsContainerImages("mcr.microsoft.com/windows/nanoserver:*", "2022-containerd")
+	serverCoreVersions := GetWindowsContainerImages("mcr.microsoft.com/windows/nanoserver:*", "2022-containerd")
 	t.Logf("found servercore version %v", serverCoreVersions)
 	require.Len(t, serverCoreVersions, 1)
 }
 
 func TestWs2022ServerCoreGen2(t *testing.T) {
-	serverCoreVersions := getWindowsContainerImages("mcr.microsoft.com/windows/servercore:*", "2022-containerd-gen2")
+	serverCoreVersions := GetWindowsContainerImages("mcr.microsoft.com/windows/servercore:*", "2022-containerd-gen2")
 	t.Logf("found servercore version %v", serverCoreVersions)
 	require.Len(t, serverCoreVersions, 2)
 }
 
 func TestWs2022NanoserverGen2(t *testing.T) {
-	serverCoreVersions := getWindowsContainerImages("mcr.microsoft.com/windows/nanoserver:*", "2022-containerd-gen2")
+	serverCoreVersions := GetWindowsContainerImages("mcr.microsoft.com/windows/nanoserver:*", "2022-containerd-gen2")
 	t.Logf("found servercore version %v", serverCoreVersions)
 	require.Len(t, serverCoreVersions, 1)
 }
 
 func TestWs2019ServerCore(t *testing.T) {
-	serverCoreVersions := getWindowsContainerImages("mcr.microsoft.com/windows/servercore:*", "2019-containerd")
+	serverCoreVersions := GetWindowsContainerImages("mcr.microsoft.com/windows/servercore:*", "2019-containerd")
 	t.Logf("found servercore version %v", serverCoreVersions)
 	require.Len(t, serverCoreVersions, 2)
 }
 
 func TestWs2019Nanoserver(t *testing.T) {
-	serverCoreVersions := getWindowsContainerImages("mcr.microsoft.com/windows/nanoserver:*", "2019-containerd")
+	serverCoreVersions := GetWindowsContainerImages("mcr.microsoft.com/windows/nanoserver:*", "2019-containerd")
 	t.Logf("found servercore version %v", serverCoreVersions)
 	require.Len(t, serverCoreVersions, 1)
 }
@@ -122,12 +122,12 @@ func TestWindowsImagesHaveServercoreAndNanoserverSpecified(t *testing.T) {
 
 	for _, image := range windowsImages {
 		t.Run(fmt.Sprintf("testing servercore has versions for %s", image.Name), func(t *testing.T) {
-			images := getServercoreImagesForVHD(image)
+			images := GetServercoreImagesForVHD(image)
 			t.Logf("found servercore version %v", images)
 			require.NotEmpty(t, images, "No Windows servercore images found")
 		})
 		t.Run(fmt.Sprintf("testing nanoserver has versions for %s", image.Name), func(t *testing.T) {
-			images := getNanoserverImagesForVhd(image)
+			images := GetNanoserverImagesForVhd(image)
 			t.Logf("found servercore version %v", images)
 			require.NotEmpty(t, images, "No Windows nanoserver images found")
 		})
